djangocms_publisher/test_project/requirements/base.txt
django-parler
